fix(monty-python): use fixed-width runtime IDs in snapshots

+++ crates/monty-python/src/monty_cls.rs
@@ -853,20 +853,23 @@ pub struct PyFunctionSnapshot {
     pub call_id: u32,
     /// Stable runtime IDs for positional args in `args` order.
     #[pyo3(get)]
-    pub arg_runtime_ids: Vec<usize>,
+    pub arg_runtime_ids: Vec<u64>,
     /// Stable runtime IDs for keyword `(key, value)` pairs in `kwargs` order.
     #[pyo3(get)]
-    pub kwarg_runtime_ids: Vec<(usize, usize)>,
+    pub kwarg_runtime_ids: Vec<(u64, u64)>,
 }
 
 /// Converts runtime-id wrappers to raw stable IDs for Python-facing snapshots.
 fn map_runtime_ids(
     arg_ids: &[::monty::RuntimeValueId],
     kwarg_ids: &[(::monty::RuntimeValueId, ::monty::RuntimeValueId)],
-) -> (Vec<usize>, Vec<(usize, usize)>) {
+) -> (Vec<u64>, Vec<(u64, u64)>) {
     (
-        arg_ids.iter().map(|id| id.raw()).collect(),
-        kwarg_ids.iter().map(|(key, value)| (key.raw(), value.raw())).collect(),
+        arg_ids.iter().map(|id| id.raw() as u64).collect(),
+        kwarg_ids
+            .iter()
+            .map(|(key, value)| (key.raw() as u64, value.raw() as u64))
+            .collect(),
     )
 }
 
@@ -1048,8 +1051,8 @@ impl PyFunctionSnapshot {
             args: Vec<MontyObject>,
             kwargs: Vec<(MontyObject, MontyObject)>,
             call_id: u32,
-            arg_runtime_ids: &'a [usize],
-            kwarg_runtime_ids: &'a [(usize, usize)],
+            arg_runtime_ids: &'a [u64],
+            kwarg_runtime_ids: &'a [(u64, u64)],
         }
 
         let snapshot = self.snapshot.lock().unwrap_or_else(PoisonError::into_inner);
@@ -1124,8 +1127,8 @@ impl PyFunctionSnapshot {
             args: Vec<MontyObject>,
             kwargs: Vec<(MontyObject, MontyObject)>,
             call_id: u32,
-            arg_runtime_ids: Vec<usize>,
-            kwarg_runtime_ids: Vec<(usize, usize)>,
+            arg_runtime_ids: Vec<u64>,
+            kwarg_runtime_ids: Vec<(u64, u64)>,
         }
 
         let bytes = data.as_bytes();
